Refactor and a couple bug fixes
Refactor a bunch of stuff to make it easier to test and separate concerns
a bit better.  Fixes #11 and #6 - we have to change how we parse the
command line.  package.variable.method and package.function are ambiguous
if the package path has a period in it (which many do).  So instad we now
will separate the package path and the function/method call with a space.
Also make sure we use the package name, not the import path in the code.